Basic uploader
Included:
imageupload.html - A basic html page with a file uploader directed at
the server. It doesn't work. I'm not sure why.

code.js - a javascript file to be used for custom scripting. Currently
it is empty.

jquery.js - The JQuery library. This has to be scripted in before any
javascript using methods from the library.