Deck Operations + JQuery UI
Create Deck from Image not yet implemented.

All other deck API is in a working state or will be as soon as server
bugs are fixed.

In addition, JQuery UI has been patched in on top of JQuery. This will
allow for some cool dynamic interface and is the first step towards a
professional-looking website. Some of it is already visible in
getdeck.html