Moving Lines Generalized
upload_image:

The uploaded image now pops up in a JQuery UI dialog box. Movable lines
now show up in the box as well. The lines and the dialog box are
tailored to fit the size of the image.

Next step: lines should actually reflect the divs I receive from the
server.